Remove TestData
