--- internal/storage/v2/clickhouse/tracestore/testdata/package_test.go
@@ -1,14 +1,0 @@
-// Copyright (c) 2025 The Jaeger Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package testdata
-
-import (
-	"testing"
-
-	"github.com/jaegertracing/jaeger/internal/testutils"
-)
-
-func TestMain(m *testing.M) {
-	testutils.VerifyGoLeaks(m)
-}
